// src/app/pages/dashboard/dashboard-request/model/service-request-line.ts
/* export interface ServiceRequestLine {
  approverAssignedBy?: string;
  approverCreatedBy?: string;
  approverCreatedDate?: string;
  approverId?: string;
  approverModifiedBy?: string;
  approverModifiedDate?: string;
  approverRemarks?: string;
  approverStatus?: string;
  channelId?: string;
  closedBy?: string;
  closedDate?: string;
  contactId?: number;
  copy?: string;
  copyFrom?: string;
  copyTo?: string;
  createdBy?: string;
  createdDate?: string;
  customerId?: number;
  priorityId?: number;
  requestId?: number;
  requestStatus?: string;
  requestType?: string;
  severityId?: number;
  updatedBy?: string;
  updatedDate?: string;
} */
export interface ServiceRequestLine {
  requestId: number;
  customerId: number;
  contactId: number;
  requestType: null;
  priorityId: number;
  severityId: number;
  channelId: null;
  createdBy: number;
  createdDate: string;
  updatedBy: null;
  updatedDate: null;
  closedBy: null;
  closedDate: null;
  requestStatus: string;
  copy: null;
  copyFrom: null;
  copyTo: null;
  approverAssignedBy: null;
  approverId: null;
  approverStatus: null;
  approverRemarks: null;
  approverCreatedDate: null;
  approverCreatedBy: null;
  approverModifiedDate: null;
  approverModifiedBy: null;
  customerName: string;
  contact: Contact;
}

export interface Contact {
  id: number;
  prefix: number;
  paxId: number;
  customerId: number;
  firstName: string;
  middleName: string;
  lastName: string;
  designationId: number;
  designationName: string;
  roleId: number;
  primaryEmail: string;
  primaryCCP: number;
  primaryPhoneNumber: number;
  secondaryEmail: string;
  secondaryCCP: number;
  secondaryPhoneNumber: number;
  telephoneNumber: number;
  remarksAndNotes: string;
  startDate: string;
  endDate: string;
  status: boolean;
  dob: string;
  passport: string;
  nationality: number;
  issuedCountry: number;
  createdBy: number;
  updatedBy: number;
  createdDate: string;
  updatedDate: string;
  dpImageUrl: string;
}